api_test_framework/requirements.txt
pytest==7.4.3
requests==2.31.0
pytest-html==4.1.1
pytest-xdist==3.3.1
python-dotenv==1.0.0
PyYAML==6.0.1 